Move saml server to test file
